Refactor MetadataGeneration
